Finalisation arène + amélioration gestion async des input players

+++ Systems/SpellSystem/Spells/RayOfFrost.cs
@@ -53,7 +53,7 @@ namespace NWN.Systems
         Task waitSpellUsed = NwTask.Run(async () =>
         {
           await NwTask.Delay(TimeSpan.FromSeconds(0.2));
-          SpellSystem.RestoreSpell(oCaster, (int)onSpellCast.Spell);
+          SpellSystem.RestoreSpell(oCaster, onSpellCast.Spell);
         });
       }      
     }
